pyproject.toml
[project]
name = "herd-ai"
version = "0.1.0"
description = "LLM Autonomous Agency — multi-agent orchestration research"
requires-python = ">=3.12"
dependencies = [
    "crewai>=0.100.0",
    "pydantic>=2.10.0",
    "fastapi>=0.115.0",
    "uvicorn>=0.34.0",
    "aiosqlite>=0.20.0",
    "structlog>=24.4.0",
    "chromadb>=0.5.0",
    "playwright>=1.49.0",
    "langfuse>=2.50.0",
    "growthbook>=1.4.0",
    "pyyaml>=6.0.2",
    "httpx>=0.28.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.3.0",
    "pytest-asyncio>=0.24.0",
    "ruff>=0.8.0",
    "mypy>=1.13.0",
]

[tool.ruff]
target-version = "py312"
line-length = 88

[tool.ruff.lint]
select = ["E", "F", "I", "N", "UP", "B", "A", "SIM", "TCH"]

[tool.ruff.lint.isort]
known-first-party = ["src"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]

[tool.mypy]
python_version = "3.12"
strict = true
warn_return_any = true
warn_unused_configs = true

[build-system]
requires = ["setuptools>=75.0"]
build-backend = "setuptools.backends._legacy:_Backend"
